requirements.txt
asgiref==3.9.1
attrs==25.3.0
django==5.2.6
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
pyjwt==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pyyaml==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing-extensions==4.15.0
uritemplate==4.2.0
